Implement evaluation functionality: 1) Added EvaluationHandler to return evaluation results  2) Added retrieval evaluation with precision, recall, f1

+++ models/retrieval.go
@@ -1,7 +1,7 @@
 package models
 
 type RetrievalResult struct {
-	ChunkID uint64
+	ChunkID int
 	Text    string
 	Score   float32 // Cosine similarity score
 }
